refactoring: moving classes adding additional interface for statistic collectors
