+++ module/Statistics/src/Calculator/StatisticCollector/AbstractCalculator.php
@@ -1,41 +1,24 @@
 <?php
 
-namespace Statistics\Calculator;
+namespace Statistics\Calculator\StatisticCollector;
 
 use SocialPost\Dto\SocialPostTo;
 use Statistics\Dto\ParamsTo;
 use Statistics\Dto\StatisticsTo;
 
-/**
- * Class AbstractCalculator
- *
- * @package Statistics\Calculator
- */
-abstract class AbstractCalculator implements CalculatorInterface
+abstract class AbstractCalculator implements StatisticCollectorInterface
 {
-
     protected const UNITS = null;
 
-    /**
-     * @var ParamsTo
-     */
-    protected $parameters;
+    protected ParamsTo $parameters;
 
-    /**
-     * @param ParamsTo $params
-     *
-     * @return CalculatorInterface
-     */
-    public function setParameters(ParamsTo $params): CalculatorInterface
+    public function setParameters(ParamsTo $params): self
     {
         $this->parameters = $params;
 
         return $this;
     }
 
-    /**
-     * @param SocialPostTo $postTo
-     */
     public function accumulateData(SocialPostTo $postTo): void
     {
         if (false === $this->checkPost($postTo)) {
@@ -45,9 +28,6 @@ abstract class AbstractCalculator implements CalculatorInterface
         $this->doAccumulate($postTo);
     }
 
-    /**
-     * @return StatisticsTo
-     */
     public function calculate(): StatisticsTo
     {
         return $this->doCalculate()
@@ -55,11 +35,6 @@ abstract class AbstractCalculator implements CalculatorInterface
                     ->setUnits(static::UNITS);
     }
 
-    /**
-     * @param SocialPostTo $postTo
-     *
-     * @return bool
-     */
     protected function checkPost(SocialPostTo $postTo): bool
     {
         if (null !== $this->parameters->getStartDate()
@@ -77,13 +52,7 @@ abstract class AbstractCalculator implements CalculatorInterface
         return true;
     }
 
-    /**
-     * @param SocialPostTo $postTo
-     */
     abstract protected function doAccumulate(SocialPostTo $postTo): void;
 
-    /**
-     * @return StatisticsTo
-     */
     abstract protected function doCalculate(): StatisticsTo;
 }
