+++ module/Statistics/src/Calculator/StatisticCollector/MaxPostLength.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Statistics\Calculator;
+namespace Statistics\Calculator\StatisticCollector;
 
 use SocialPost\Dto\SocialPostTo;
 use Statistics\Dto\StatisticsTo;
