+++ module/Statistics/src/Calculator/StatisticCollector/StatisticCollectorInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Statistics\Calculator\StatisticCollector;
+
+use SocialPost\Dto\SocialPostTo;
+use Statistics\Dto\StatisticsTo;
+
+interface StatisticCollectorInterface
+{
+    public function accumulateData(SocialPostTo $postTo): void;
+
+    public function calculate(): StatisticsTo;
+}
